% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{ClipValues}
\alias{ClipValues}
\title{ClipValues}
\usage{
ClipValues(values, max.thresh, min.thresh = 0)
}
\arguments{
\item{values}{(vector) input variables to clip.}

\item{max.thresh}{(numeric) values above this value will be assigned to this value.}

\item{min.thresh}{(numeric) values below this value will be assigned to this value.}
}
\value{
values input vector, altered to have new minimum/maximum values.
}
\description{
\code{ClipValues} Clips the extremely large or small values of an input vector.
}
\details{
Clips the extremely large or small values of an input vector.
}
\examples{

}
