% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{formatGraphTitleForFileName}
\alias{formatGraphTitleForFileName}
\title{formatGraphTitleForFileName}
\usage{
formatGraphTitleForFileName(title)
}
\arguments{
\item{-}{title: character string to remove characters of}
}
\value{

}
\description{
\code{formatGraphTitleForFileName} Removes illegal characters for titles so they can be used for saving files
}
\details{
Removes illegal characters for titles so they can be used for saving files
}
\examples{

}
