% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateaWhereChart.R
\name{generateaWhereChart}
\alias{generateaWhereChart}
\title{generateaWhereChart}
\usage{
generateaWhereChart(
  data,
  variable,
  variable_rightAxis = NULL,
  title = NULL,
  e_precip = FALSE,
  e_threshold = 35,
  doRoll = FALSE,
  rolling_window = 30,
  includeSTD = FALSE,
  mainGraphType = "line",
  daysToAggregateOver = NULL,
  yAxisLimits = NA
)
}
\arguments{
\item{-}{yAxisLimits: Used to set the limits of the y axis explicitly.  If used, must be a two element vector of the form 
c(minValue, maxValue) (optional)}
}
\value{
plot object
}
\description{
\code{generateaWhereChart} Generate a plot using aWhere weather data with standardized formatting.
}
\details{
This function makes basic line plots using data returned from the aWhereAPI and formatted using
the generateaWhereDataset function. The function requires variables to be named according to the
conventions of that function, or the basic functions of the aWhereAPI package. Users must pass the dataset object
to the function as well as one of the acceptable list of variables to chart. Users may also
use the optional parameters to customize a plot title, turn on the "effective precipitation"
option, or customize the rolling window. The returned plot will be a line chart comparing
the actual conditions within the dataset's datespan (labeled Current in the plot) to the 
long-term normals for that same datespan.

If users turn on effective precipitation (e_precip = TRUE), the function will add a line to 
the plot in which all daily precipitation values (the Current line) are capped at the maximum 
allowable amount (defined by the parameter e_threshold). This serves as a proxy for runoff
in times of high volumes of rain. The default value of e_threshold is 35 mm.

The standard plot formatting is generated using ggplot2's gray theme.
}
\examples{
\dontrun{generateaWhereChart(data = weather_df
                             ,variable = "accumulatedPrecipitation" 
                             ,e_precip = TRUE
                             ,e_threshold = 10
                             ,doRoll = TRUE)}
}
\references{
http://developer.awhere.com/api/reference/
}
