% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{generateColorScale}
\alias{generateColorScale}
\title{generateColorScale}
\usage{
generateColorScale(fig, add.years, add.years.colors)
}
\arguments{
\item{fig}{ggplot figure}

\item{add.years}{years to add to plot}

\item{add.years.colors}{explicit colors for each year}
}
\value{

}
\description{
\code{generateColorScale} #used to generate colorscheme when additional years are added to graph
}
\details{
#used to generate colorscheme when additional years are added to graph
}
\examples{

}
