% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoordinateInfo.R
\name{getCoordinateInfo}
\alias{getCoordinateInfo}
\title{getCoordinateInfo}
\usage{
getCoordinateInfo(points, attribute)
}
\arguments{
\item{-}{points: data frame or matrix of points, with longitude in the first column and
latitude in the second. Points should be in decimal degrees. (required)}

\item{-}{attribute: character string of the type of information desired, accepted inputs 
are "country", "continent", or "continent classification". (required)}
}
\value{
vector of names with factor levels
}
\description{
\code{getCoordinateInfo} Find some basic information about the geography in which a lat/long point falls.
}
\details{
This function returns the name of the country in which a given point is located,
or if desired, the continent, the development designation, or the population of 
the country. Users of the aWhere API often pull weather data for lists of  
points or fields for which no other geographic information is provided, or receive 
prepackaged large weather data files identified by points which do not include 
other geographic information for size reasons. This function helps to retrieve this
information quickly if needed for analysis without requiring modification of the 
original files.
}
\examples{
\dontrun{getCoordinateInfo(points = weather_df[, c("longitude", "latitude")], attribute = "country")}
}
\references{

}
