% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{GetDays}
\alias{GetDays}
\title{getDays}
\usage{
GetDays(starting.day, forecast.days, ending.day = FALSE)
}
\arguments{
\item{-}{data: data frame in which variables are named according to the schema output by generateaWhereChart.R (required)}

\item{-}{starting.day: (string) the first day of the forecast. Acceptable values are "today","yesterday", "tomorrow",
or a specific date with format "YYYY-MM-DD"}

\item{-}{forecast.days: (vector of numeric value(s)) duration of the forecast(s). For a 7- and 3-day forecast: c(7,3).}

\item{-}{ending.day: (optional, string with format "YYYY-MM-DD") if supplied, this will be the last day of the forecast.
Otherwise, the ending day is calculated by adding the greatest value in forecast.days to the starting date.}
}
\value{
# days (list of characters) starting day in position 1; ending day in position 2
}
\description{
\code{getDays} Generate the starting and ending days for a forecast.
}
\details{
Generate the starting and ending days for a forecast.
}
\examples{

}
