% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{GetForecastData}
\alias{GetForecastData}
\title{getForecastData}
\usage{
GetForecastData(template.place, days, years, write.file = FALSE,
  filename.out = "forecast")
}
\arguments{
\item{-}{template.place: (data.frame) contains the geographic data for each grid cell}

\item{-}{days: (list of characters) starting day in position 1; ending day in position 2
each with the format "YYYY-MM-DD"}

\item{-}{years: (list of numerics) starting year in position 1; ending year in position 2}

\item{-}{write.file: (optional, boolean with default value FALSE) when true, the forecast data is written to a .csv file}

\item{-}{filename.out (optional, character string with default value "forecast") used to name the output file}
}
\value{
forecast.all: (data.frame) containing all of the n-day forecast data.
}
\description{
\code{getForecastData} pulls forecast data from the aWhere API for the specified days/years.
}
\details{
pulls forecast data from the aWhere API for the specified days/years.
}
\examples{

}
