% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{GetForecastSummary}
\alias{GetForecastSummary}
\title{GetForecastSummary}
\usage{
GetForecastSummary(forecast.all, n.day.forecasts, template.place)
}
\arguments{
\item{-}{forecast.all: (data.frame) containing all of the n-day forecast data.}

\item{-}{n.day.forecasts (vector of numeric value(s)) duration of the forecast(s).
For a 7- and 3-day forecast: c(7,3).}

\item{-}{template.place (data.frame) contains the geographic data for each grid cell}
}
\value{
forecast.all: (data.frame) containing all of the n-day forecast data.
}
\description{
\code{GetForecastSummary} Sum the n-day forecast values at every location.
}
\details{
Sum the n-day forecast values at every location.
}
\examples{

}
