% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{loadVarMemRef_aWhereEnv}
\alias{loadVarMemRef_aWhereEnv}
\title{loadVarMemRef_aWhereEnv}
\usage{
loadVarMemRef_aWhereEnv(memAddress)
}
\arguments{
\item{memAddress}{address of object in memory}
}
\value{

}
\description{
\code{loadVarMemRef_aWhereEnv} load the data object stored at a specific memory address
}
\details{
load the data object stored at a specific memory address
}
\examples{

}
