% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClimateTrendTable.R
\name{makeClimateTrendTable}
\alias{makeClimateTrendTable}
\title{makeClimateTrendTable}
\usage{
makeClimateTrendTable(
  data,
  filename,
  variable,
  season.monthDay_start = "01-01",
  season.monthDay_end = "12-31",
  years.LTN = seq(2006, 2020, 1),
  title = NULL,
  e_precip = FALSE,
  e_threshold = 35,
  writeOutTableData = TRUE,
  indexSpecificValue = NULL
)
}
\arguments{
\item{data}{data frame in which variables are named according to the schema
output by generateaWhereDataset.R (required)}

\item{filename}{filename for output image.  Periods are not allowed except before file
type (required)}

\item{variable}{character string denoting the variable to chart. Acceptable
values are maxLenDrySpell, maxLenWetSpell, numFrostDays, numSummerDays, numIcingDays,
numTropicalNights, minOfMaxTemp, maxOfMaxTemp, minOfMinTemp, maxOfMinTemp,
dailyTempRange, maxSingleDayPrecip, max5ConsDayPrecip, simplePrecipIntensityIndex,
precipSumExceedPercentile, warmSpellDurIndex, coldSpellDurIndex,
countDaysPrecipExceedAmount, percentDaysMinTempBelowQuantile, percentDaysMaxTempBelowQuantile,
percentDaysMinTempAboveQuantile, percentDaysMaxTempAboveQuantileaccumulatedGdd, sumOfGdd, sumOfPET,
sumOfPrecip, sumOfSolar, averageMaxTemp, averageMinTemp, averageMaxRH, averageMinRH,
averageWind, and maxWindGusts}

\item{season.monthDay_start}{Specify the start month-day combination of the "season"
you want analyzed (optional)}

\item{season.monthDay_end}{Specify the end month-day combination of the "season"
you want analyzed (optional)}

\item{years.LTN}{specify the years over which you want the LTN calculated.  Defaults
to 2006-2020 (optional)}

\item{title}{character string of title to assign to the plot. (optional)}

\item{e_precip}{logical, if set to TRUE, effective precipitation will be
calculated and charted based on e_threshold. Default is set to FALSE.
(optional)}

\item{e_threshold}{numeric value (in milimeters) for the daily maximum used
to calculate effective precipitation if e_precip is set to TRUE. (optional)}

\item{writeOutTableData}{Boolean for whether the underlying data in the table
should be written to CSV.  File name will match the value of the "filename"
parameter with .csv file type (optional)}

\item{indexSpecificValue}{For the Climate Indices this tool can plot the user
can override the default value of the index using this parameter (optional)}
}
\value{
plot object
}
\description{
\code{makeClimateTrendTable} Generate a table across multiple growing seasons
using aWhere weather data with standardized formatting of a prefedined set of climate indices
}
\details{
This function extends the aWhere charts package to make a table with the ETCCDI climate
change indices (reference below).  Most indices are implemented with the default behavior 
described int he publication.  Default values of the indices can be overriden using the 
"indexSpecificValue" parameter. Applies a summary statistic over each year/season of data

Also has helper functions including to automatically (with user permission) download missing
but needed data and store in the aWhereEnv to persist it throughout a user R session.
}
\examples{
\dontrun{makeClimateTrendTable(data = weather_df
                               ,filename = 'test.png'
                               ,variable = "seasonTotalPrecip"
                               ,season.monthDay_start = '09-01'
                               ,season.monthDay_end = '11-30'
                               ,years.LTN = seq(2010,2019,1)
                               ,e_precip = TRUE
                               ,e_threshold = 10)}

}
\references{
http://etccdi.pacificclimate.org/list_27_indices.shtml
}
