% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{MakeMap}
\alias{MakeMap}
\title{MakeMap}
\usage{
MakeMap(df, v = "pre", base.map, map.name, write.file = FALSE)
}
\arguments{
\item{-}{df Data frame ("df") containing CLIMARK data}

\item{-}{v: Character string specifying the variable to map. Default "pre".
Acceptable values for "v" are: 
     "pre" = precipitation
     "ppet" = P / PET 
     "pltn" = recip vrs. LTN precip}

\item{-}{base.map terrain map image from Google Maps to overlay the mapped variable}

\item{-}{map.name: character string to name the map and output accordingly}

\item{-}{write.file: (optional, boolean with default value FALSE) when true, the
forecast data is written to a .csv file}
}
\value{
values input vector, altered to have new minimum/maximum values.
}
\description{
\code{MakeMap} this function creates a climark map using CLIMARK data
}
\details{
this function creates a climark map using CLIMARK data
}
\examples{

}
