% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{MapForecast}
\alias{MapForecast}
\title{MapForecast}
\usage{
MapForecast(forecasts.n, base.map, thresholds)
}
\arguments{
\item{-}{forecasts.n (data frame) contains the columns "n.day" (the number of days
used to aggregate forecast data), precip.amount.sum (aggregated)
precipitation amount), and shapewkt (polygon data), the data to be 
mapped.}

\item{-}{base.map (ggmap raster) map object on which to overlay the gridded weather data}

\item{-}{thresholds (data frame) data frame containing the minimum (first row) and maximum 
(second row) threshold values to clip weather variables (column names)}
}
\value{

}
\description{
\code{MapForecast} CMaps the Forecast
}
\details{
Calls the MakeMap function to map forecasted weather data aggregated by n-days
(as specified in the forecasts.n data frame). Current functionality only supports
precipitation data, but can easily be modified to map other variables as well.
}
\examples{

}
