% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnAppropriateSummaryStatistic.R
\name{returnAppropriateSummaryStatistic}
\alias{returnAppropriateSummaryStatistic}
\title{returnAppropriateSummaryStatistic}
\usage{
returnAppropriateSummaryStatistic(variable)
}
\arguments{
\item{variable}{character string denoting the variable to chart. Acceptable
values are maxLenDrySpell, maxLenWetSpell, numFrostDays, numSummerDays, numIcingDays,
numTropicalNights, minOfMaxTemp, maxOfMaxTemp, minOfMinTemp, maxOfMinTemp,
dailyTempRange, maxSingleDayPrecip, max5ConsDayPrecip, simplePrecipIntensityIndex,
precipSumExceedPercentile, warmSpellDurIndex, coldSpellDurIndex,
countDaysPrecipExceedAmount, percentDaysMinTempBelowQuantile, percentDaysMaxTempBelowQuantile,
percentDaysMinTempAboveQuantile, percentDaysMaxTempAboveQuantile}
}
\value{
string
}
\description{
\code{returnAppropriateSummaryStatistic} Internal function to return
appropriate summary statistic to use with each index
}
\details{
Internal function to return
appropriate summary statistic to use with each index
}
