% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{WriteJpeg}
\alias{WriteJpeg}
\title{WriteJpeg}
\usage{
WriteJpeg(plt, plt.title, w = 10, h = 6, u = "in", r = 500)
}
\arguments{
\item{-}{r (integer, optional with default value of 500) nominal resolution in ppi}
}
\value{

}
\description{
\code{WriteJpeg} Saves the plot as a jpeg
}
\details{
Opens the jpeg graphics device with the specified settings and saves the plot
to an image file (.jpeg)
}
\examples{

}
