% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateaWhereChart.R
\name{generateaWhereChart}
\alias{generateaWhereChart}
\title{generateaWhereChart}
\usage{
generateaWhereChart(data, variable, title, e_precip = FALSE,
  e_threshold = 35, rolling_window = 30)
}
\arguments{
\item{-}{data: data frame in which variables are named according to the schema output by generateaWhereChart.R (required)}

\item{-}{variable: character string denoting the variable to chart. Acceptable values 
are precipitation, accumulatedPrecipitation, maxTemp, minTemp, pet, accumulatedPet, ppet, 
or rollingavgppet. (required)}

\item{-}{title: character string of title to assign to the plot. (required)}

\item{-}{e_precip: logical, if set to TRUE, effective precipitation will 
be calculated and charted based on e_threshold. Default is set to FALSE. (optional)}

\item{-}{e_threshold: numeric value (in milimeters) for the daily maximum used to calculate 
effective precipitation if e_precip is set to TRUE. (optional)}

\item{-}{rolling_window: numeric value for the number of days to use in rolling 
average calculations, only applicable if the variable parameter is set to
"rollingavgppet". Default value is 30. (optional)}
}
\value{
plot object
}
\description{
\code{generateaWhereChart} Generate a plot using aWhere weather data with standardized formatting.
}
\details{
This function makes basic line plots using data returned from the aWhereAPI and formatted using
the generateaWhereDataset function. The function requires variables to be named according to the
conventions of that function, or the basic functions of the aWhereAPI package. Users must pass the dataset object
to the function as well as one of the acceptable list of variables to chart. Users may also
use the optional parameters to customize a plot title, turn on the "effective precipitation"
option, or customize the rolling window. The returned plot will be a line chart comparing
the actual conditions within the dataset's datespan (labeled Current in the plot) to the 
long-term normals for that same datespan.

If users turn on effective precipitation (e_precip = TRUE), the function will add a line to 
the plot in which all daily precipitation values (the Current line) are capped at the maximum 
allowable amount (defined by the parameter e_threshold). This serves as a proxy for runoff
in times of high volumes of rain. The default value of e_threshold is 35 mm.

The standard plot formatting is generated using ggplot2's gray theme.
}
\examples{
\dontrun{generateaWhereChart(data = weather_df, variable = "accumulatedPrecipitation", 
                             title = "Accumulated Precipitation values at Location from Xdate to Ydate",
                             e_precip = TRUE, e_threshold = 20)}
}
\references{
http://developer.awhere.com/api/reference/
}
