% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateaWhereChart.R
\name{generateaWhereChart}
\alias{generateaWhereChart}
\title{generateaWhereChart}
\usage{
generateaWhereChart(
  data,
  variable,
  variable_rightAxis = NULL,
  day_start = NULL,
  day_end = NULL,
  title = NULL,
  e_precip = FALSE,
  e_threshold = 35,
  doRoll = FALSE,
  rolling_window = 30,
  includeSTD = FALSE,
  mainGraphType = "line",
  daysToAggregateOver = NULL,
  yAxisLimits = NA,
  size_font_main_title = 16,
  size_font_axis_titles = 14,
  size_font_axis_labels = 12,
  size_font_legend_entries = 12,
  line_width = 1,
  annotationsWhichSide = "left",
  indexSpecificValue = NULL
)
}
\arguments{
\item{data}{data frame in which variables are named according to the schema
output by generateaWhereDataset.R (required)}

\item{variable}{character string denoting the variable to chart. Acceptable
values are accumulatedGdd, accumulatedPet, accumulatedPpet,
accumulatedPrecipitation, gdd, pet, precipitation, maxRH, minRH,
solar,averageWind,dayMaxWind, rollingavgppet. (required)}

\item{variable_rightAxis}{What variable to plot over the primary variable.
The right y-axis of the plot will be used to present its range. Note that
it will always be plotted as a line chart. Same valid values as the
variable param.  (optional)}

\item{day_start}{Specify the start date of the daily data you want plotted in
the figure (optional)}

\item{day_end}{Specify the end date of the daily data you want plotted in the
figure (optional)}

\item{title}{character string of title to assign to the plot. (required)}

\item{e_precip}{logical, if set to TRUE, effective precipitation will be
calculated and charted based on e_threshold. Default is set to FALSE.
(optional)}

\item{e_threshold}{numeric value (in milimeters) for the daily maximum used
to calculate effective precipitation if e_precip is set to TRUE. (optional)}

\item{doRoll}{apply a rolling average to the calculation.}

\item{rolling_window}{numeric value for the number of days to use in rolling
average calculations.  Default value is 30. (optional)}

\item{includeSTD}{whether to plot the standard deviation as a ribbon around
the LTN value of the main variable. (optional)}

\item{daysToAggregateOver}{Used to temporally aggregate data.  Unit is in
days. This is done based on the startdate of the dataset, not a calendar
week (otpional)}

\item{yAxisLimits}{Used to set the limits of the y axis explicitly.  If used,
must be a two element vector of the form c(minValue, maxValue) (optional)}

\item{size_font_main_title}{Font size of main title of graph (optional)}

\item{size_font_axis_titles}{Font size of axes on graph (optional)}

\item{size_font_axis_labels}{Font size of labels on axes on graph (optional)}

\item{size_font_legend_entries}{Font size of entries in lengend on graph (optional)}

\item{line_width}{Font size for line geometries on charts (optional)}

\item{indexSpecificValue}{For the Climate Indices this tool can plot the user
can ovveride the default value of the index using this parameter (optional)}

\item{maingraphType}{Which type of graph to make for the main plot.  Valid
values are "line" and "bar" (optional)}
}
\value{
plot object
}
\description{
\code{generateaWhereChart} Generate a plot using aWhere weather
data with standardized formatting.
}
\details{
This function makes basic line plots using data returned from the
aWhereAPI and formatted using the generateaWhereDataset function. The
function requires variables to be named according to the conventions of that
function, or the basic functions of the aWhereAPI package. Users must pass
the dataset object to the function as well as one of the acceptable list of
variables to chart. Users may also use the optional parameters to customize a
plot title, turn on the "effective precipitation" option, or customize the
rolling window. The returned plot will be a line chart comparing the actual
conditions within the dataset's datespan (labeled Current in the plot) to the
long-term normals for that same datespan.

If users turn on effective precipitation (e_precip = TRUE), the function will
add a line to the plot in which all daily precipitation values (the Current
line) are capped at the maximum allowable amount (defined by the parameter
e_threshold). This serves as a proxy for runoff in times of high volumes of
rain. The default value of e_threshold is 35 mm.

The standard plot formatting is generated using ggplot2's gray theme.
}
\examples{
\dontrun{generateaWhereChart(data = weather_df
                             ,variable = "accumulatedPrecipitation"
                             ,e_precip = TRUE
                             ,e_threshold = 10
                             ,doRoll = TRUE)}

}
\references{
http://developer.awhere.com/api/reference/
}
