% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMultiplot.R
\name{generateMultiplot}
\alias{generateMultiplot}
\title{generateMultiplot}
\usage{
generateMultiplot(..., title, fontsize = 12, cols = 2)
}
\arguments{
\item{-}{...: All plot objects to assemble, separated by commas. (required)}

\item{-}{title: character string of the title you would like to assign to the multiplot,
multiplot title is independent of any titles set to individual plots. (required)}

\item{-}{fontsize: numeric value of the fontsize to use in rendering the multiplot title, default is 12. (optional)}

\item{-}{cols: numeric value for the number of columns into which to assemble input plots, default is 2. (optional)}
}
\value{
plot object
}
\description{
\code{generateMultiplot} Assembles multiple plots into a single window
}
\details{
This function allows you to assemble several plot objects into a single window
for more convenient display and saving. The function has options for customizing 
the number of columns (into which plots are arranged), and the title assigned
to the overall multiplot. This function was created to serve users who were generating 
multiple weather/agronomic plots using functions elsewhere in this package, but
wanted to be able to combine them into a single "analysis" window in order to more 
easily detect combinations of conditions.
}
\examples{
\dontrun{generateMultiplot(rain_plot, maxtemp_plot, fontsize = 20, cols = 1)}
}
\references{
See other functions in this package for generating plots as inputs.
}
